mcri_deploy/docker/redis-cache-obs/requirements.txt
requests==2.31.0
tqdm==4.65.0
redis==4.6.0
docopt==0.6.2
bgzip==0.4.0
google-cloud-storage==2.14.0
